Dice Roller (Dungeons and Dragons format)
The input is one or more lines in the format of, for example, 3d6. This means 3 dice are being rolled and each has 6 sides. The number of dice can vary from 1-100 and the number of sides on the dice can vary from 2-100. Returns the value of each roll as well as the sum of the input.
This is an implementation of one dice input.
Next will be multiple inputs.